Add working version of getRecipe of CalculatorServiceImpl
Change quantity from int to long to relay with order book service responses, add convenient fields to ProductDto

+++ calculator/src/main/java/com/market/banica/calculator/model/Product.java
@@ -1,13 +1,9 @@
 package com.market.banica.calculator.model;
 
 import com.market.banica.calculator.enums.UnitOfMeasure;
-import lombok.AllArgsConstructor;
 import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Component;
 
-import java.io.Serializable;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -19,5 +15,5 @@ public class Product  {
 
     private UnitOfMeasure unitOfMeasure;
 
-    private Map<String,Integer> ingredients = new HashMap<>();
+    private Map<String,Long> ingredients = new HashMap<>();
 }
